refactor: split the story drawer's contract out of its view
Adding the STATUS field pushed issue-detail-dialog.tsx to 510 lines, past the
cap. The seam was already there: the top ~60 lines are what the drawer
PROMISES the kanban page — its options and the shape of every way it can be
left — and the rest is how it draws itself.

Those halves change for different reasons. The contract is what the page
switches on and what the render tests assert a keypress produced; the view is
fields, focus cycle, paste handling, chip rows. Splitting them means a change
to the drawer's layout cannot quietly change what the page receives, and the
page can name the contract without pulling in the view.

The dialog re-exports both types, so callers and tests keep one import site.
472 lines and 60.
