+++ packages/kobe/src/tui-react/component/issue-detail-dialog.tsx
@@ -50,50 +50,12 @@ import {
   DialogLabel,
   DialogSection,
 } from "../ui/dialog-parts"
+import type { IssueDetailOptions, IssueDetailOutcome, IssueDraft } from "./issue-detail-contract"
 import { IssueEventsSection } from "./issue-detail-parts"
 
-export interface IssueDetailOptions {
-  readonly issue: Issue
-  /** `create` turns the drawer into the new-story intake: blank drafts,
-   *  esc CANCELS (nothing exists to save), ctrl+s creates without starting,
-   *  enter/ctrl+enter creates AND starts at the chosen placement. */
-  readonly mode?: "detail" | "create"
-  /** Engines to offer (detected built-ins + custom), in cycle order. */
-  readonly engines: readonly VendorId[]
-  readonly defaultVendor: VendorId
-  readonly engineLabel: (vendor: VendorId) => string
-  /** Live daemon connection — the linked story's EVENTS feed reads from it.
-   *  Absent (mocks, offline): the feed renders its empty state. */
-  readonly orchestrator?: RemoteOrchestrator | null
-}
-
-/** The drafted title/body/status every non-create outcome carries — the page
- *  saves a dirty patch regardless of how the drawer was left. */
-interface IssueDraft {
-  readonly title: string
-  readonly body: string
-  readonly status: IssueStatus
-}
-
-/** Every outcome carries the drafted title/body/status — the page saves a
- *  dirty patch regardless of how the drawer was left. `jump` is the drawer's
- *  follow-or-stay toggle, orthogonal to placement. */
-export type IssueDetailOutcome =
-  | ({ kind: "start"; vendor: VendorId; placement: IssueChatPlacement; jump: boolean } & IssueDraft)
-  | ({ kind: "open"; taskId: string } & IssueDraft)
-  /** Drop the story's task link — the only way back out of In progress when
-   *  the linked task is gone (deleted before the daemon unlinked, or a store
-   *  restored from an older home). The page clears `taskId`; the task, its
-   *  branch and its worktree are untouched. */
-  | ({ kind: "unlink" } & IssueDraft)
-  | ({ kind: "close" } & IssueDraft)
-  /** Create-mode result — `start` null = save only ("New story" Save). */
-  | {
-      kind: "create"
-      title: string
-      body: string
-      start: { vendor: VendorId; placement: IssueChatPlacement; jump: boolean } | null
-    }
+// The contract lives next door; re-exported here so the page and the render
+// tests keep one import site for "the drawer and what it answers with".
+export type { IssueDetailOptions, IssueDetailOutcome } from "./issue-detail-contract"
 
 type Field = "title" | "description" | "status" | "engine" | "workspace" | "jump" | "open" | "unlink"
